Wait for refill to end before writing into bufram
When doing burst writes, and the writes matches the address
that's in the bufram, it's not enough to initially check that
the first write address is "clean" (i.e. has been written into
the bufram from the read side, in case of a refill).
Instead check that all of the slots in the bufram is clean before
performing the write into it.