wb_port: fix read-after-write bug
if a read to an address already in the burst cache
would immediately occur after a write to the same
burst cache, the address from the write would be
used to fetch the read data.

This fixes it by disabling support for write bursts
(which were broken in either case) and deassert the
write to the burst cache when the write is acked.